+++ besu/src/main/java/org/hyperledger/besu/Runner.java
@@ -140,10 +140,12 @@ public class Runner implements AutoCloseable {
       transactionPoolEvictionService.start();
 
       LOG.info("Ethereum main loop is up.");
+      // we write these values to disk to be able to access them during the acceptance tests
       writeBesuPortsToFile();
       writeBesuNetworksToFile();
       writePidFile();
     } catch (final Exception ex) {
+      LOG.error("unable to start main loop", ex);
       throw new IllegalStateException("Startup failed", ex);
     }
   }
@@ -267,6 +269,10 @@ public class Runner implements AutoCloseable {
     if (port.isPresent()) {
       properties.setProperty("metrics", String.valueOf(port.get()));
     }
+    port = getEngineJsonRpcPort();
+    if (port.isPresent()) {
+      properties.setProperty("engine-json-rpc", String.valueOf(port.get()));
+    }
     // create besu.ports file
     createBesuFile(
         properties, "ports", "This file contains the ports used by the running instance of Besu");
@@ -358,7 +364,6 @@ public class Runner implements AutoCloseable {
       final Properties properties, final String fileName, final String fileHeader) {
     final File file = new File(dataDir.toFile(), String.format("besu.%s", fileName));
     file.deleteOnExit();
-
     try (final FileOutputStream fileOutputStream = new FileOutputStream(file)) {
       properties.store(
           fileOutputStream,