+++ ethereum/api/src/main/java/org/hyperledger/besu/ethereum/api/jsonrpc/internal/methods/TraceCall.java
@@ -19,12 +19,9 @@ import static org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRp
 
 import org.hyperledger.besu.ethereum.api.jsonrpc.RpcMethod;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.JsonRpcRequestContext;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.BlockParameter;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.parameters.TraceTypeParameter;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.processor.TransactionTrace;
 import org.hyperledger.besu.ethereum.api.jsonrpc.internal.response.JsonRpcErrorResponse;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.flat.FlatTrace;
-import org.hyperledger.besu.ethereum.api.jsonrpc.internal.results.tracing.flat.MixInIgnoreRevertReason;
 import org.hyperledger.besu.ethereum.api.query.BlockchainQueries;
 import org.hyperledger.besu.ethereum.core.Block;
 import org.hyperledger.besu.ethereum.core.BlockHeader;
@@ -36,31 +33,22 @@ import org.hyperledger.besu.ethereum.vm.DebugOperationTracer;
 import java.util.Optional;
 import java.util.Set;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
 public class TraceCall extends AbstractTraceByBlock implements JsonRpcMethod {
+  private static final Logger LOG = LoggerFactory.getLogger(TraceCall.class);
+
   public TraceCall(
       final BlockchainQueries blockchainQueries,
       final ProtocolSchedule protocolSchedule,
       final TransactionSimulator transactionSimulator) {
     super(blockchainQueries, protocolSchedule, transactionSimulator);
-    // The trace_call specification does not output the revert reason, so we have to remove it
-    mapper.addMixIn(FlatTrace.class, MixInIgnoreRevertReason.class);
   }
 
   @Override
   public String getName() {
     return transactionSimulator != null ? RpcMethod.TRACE_CALL.getMethodName() : null;
-  }
-
-  @Override
-  protected BlockParameter blockParameter(final JsonRpcRequestContext request) {
-    final Optional<BlockParameter> maybeBlockParameter =
-        request.getOptionalParameter(2, BlockParameter.class);
-
-    if (maybeBlockParameter.isPresent()) {
-      return maybeBlockParameter.get();
-    }
-
-    return BlockParameter.LATEST;
   }
 
   @Override
@@ -70,7 +58,7 @@ public class TraceCall extends AbstractTraceByBlock implements JsonRpcMethod {
         requestContext.getRequiredParameter(1, TraceTypeParameter.class);
 
     final Optional<BlockHeader> maybeBlockHeader =
-        blockchainQueries.get().getBlockHeaderByNumber(blockNumber);
+        blockchainQueriesSupplier.get().getBlockHeaderByNumber(blockNumber);
 
     if (maybeBlockHeader.isEmpty()) {
       return new JsonRpcErrorResponse(requestContext.getRequest().getId(), BLOCK_NOT_FOUND);
@@ -86,16 +74,24 @@ public class TraceCall extends AbstractTraceByBlock implements JsonRpcMethod {
             maybeBlockHeader.get());
 
     if (maybeSimulatorResult.isEmpty()) {
+      LOG.error(
+          "Empty simulator result, call params: {}, blockHeader: {} ",
+          JsonCallParameterUtil.validateAndGetCallParams(requestContext),
+          maybeBlockHeader.get());
+      return new JsonRpcErrorResponse(requestContext.getRequest().getId(), INTERNAL_ERROR);
+    }
+    final TransactionSimulatorResult simulatorResult = maybeSimulatorResult.get();
+
+    if (simulatorResult.isInvalid()) {
+      LOG.error(String.format("Invalid simulator result %s", maybeSimulatorResult));
       return new JsonRpcErrorResponse(requestContext.getRequest().getId(), INTERNAL_ERROR);
     }
 
     final TransactionTrace transactionTrace =
         new TransactionTrace(
-            maybeSimulatorResult.get().getTransaction(),
-            maybeSimulatorResult.get().getResult(),
-            tracer.getTraceFrames());
+            simulatorResult.getTransaction(), simulatorResult.getResult(), tracer.getTraceFrames());
 
-    final Block block = blockchainQueries.get().getBlockchain().getChainHeadBlock();
+    final Block block = blockchainQueriesSupplier.get().getBlockchain().getChainHeadBlock();
 
     return getTraceCallResult(
         protocolSchedule, traceTypes, maybeSimulatorResult, transactionTrace, block);