+++ ethereum/api/src/test/java/org/hyperledger/besu/ethereum/api/jsonrpc/AbstractJsonRpcHttpBySpecTest.java
@@ -30,6 +30,8 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 import java.util.stream.Stream;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
@@ -50,6 +52,13 @@ import org.junit.runners.Parameterized;
 public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpServiceTest {
 
   private static final ObjectMapper objectMapper = new ObjectMapper();
+  private static final Pattern GAS_MATCH_FOR_STATE_DIFF =
+      Pattern.compile("\"balance\":(?!\"=\").*?},");
+  private static final Pattern GAS_MATCH_FOR_VM_TRACE =
+      Pattern.compile(",*\"cost\":[0-9a-fA-F]+,*|,\"used\":[0-9a-fA-F]+");
+  private static final Pattern GAS_MATCH_FOR_TRACE =
+      Pattern.compile("\"gasUsed\":\"[x0-9a-fA-F]+\",");
+
   private final URL specURL;
 
   protected AbstractJsonRpcHttpBySpecTest(final String specName, final URL specURL) {
@@ -113,6 +122,7 @@ public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpS
     final String json = Resources.toString(specFile, Charsets.UTF_8);
     final ObjectNode specNode = (ObjectNode) objectMapper.readTree(json);
     final String rawRequestBody = specNode.get("request").toString();
+
     final RequestBody requestBody = RequestBody.create(JSON, rawRequestBody);
     final Request request = new Request.Builder().post(requestBody).url(baseUrl).build();
 
@@ -125,7 +135,11 @@ public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpS
         try {
           final ObjectNode responseBody =
               (ObjectNode) objectMapper.readTree(Objects.requireNonNull(resp.body()).string());
-          checkResponse(responseBody, (ObjectNode) expectedResponse);
+          checkResponse(
+              responseBody,
+              (ObjectNode) expectedResponse,
+              getMethod(rawRequestBody),
+              getTraceType(specFile.toString()));
         } catch (final Exception e) {
           throw new RuntimeException("Unable to parse response as json object", e);
         }
@@ -139,13 +153,52 @@ public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpS
         }
         for (int i = 0; i < ((ArrayNode) expectedResponse).size(); i++) {
           checkResponse(
-              (ObjectNode) responseBody.get(i), (ObjectNode) ((ArrayNode) expectedResponse).get(i));
+              (ObjectNode) responseBody.get(i),
+              (ObjectNode) ((ArrayNode) expectedResponse).get(i),
+              getMethod(rawRequestBody),
+              getTraceType(specFile.toString()));
         }
       }
     }
   }
 
-  private void checkResponse(final ObjectNode responseBody, final ObjectNode expectedResponse)
+  private enum Method {
+    TRACE_CALL_MANY,
+    OTHER
+  }
+
+  private enum TraceType {
+    TRACE,
+    VM_TRACE,
+    STATE_DIFF,
+    OTHER
+  }
+
+  private Method getMethod(final String rawRequest) {
+    if (rawRequest.contains("\"method\":\"trace_callMany\"")) {
+      return Method.TRACE_CALL_MANY;
+    } else {
+      return Method.OTHER;
+    }
+  }
+
+  private TraceType getTraceType(final String fileName) {
+    if (fileName.contains("_trace")) {
+      return TraceType.TRACE;
+    } else if (fileName.contains("_vmTrace")) {
+      return TraceType.VM_TRACE;
+    } else if (fileName.contains("_stateDiff")) {
+      return TraceType.STATE_DIFF;
+    } else {
+      return TraceType.OTHER;
+    }
+  }
+
+  private void checkResponse(
+      final ObjectNode responseBody,
+      final ObjectNode expectedResponse,
+      final Method method,
+      final TraceType traceType)
       throws JsonProcessingException {
     // Check id
     final String actualId = responseBody.get("id").toString();
@@ -160,17 +213,48 @@ public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpS
     // Check result
     if (expectedResponse.has("result")) {
       assertThat(responseBody.has("result")).isTrue();
-      final String expectedResult = expectedResponse.get("result").toString();
-      final String actualResult = responseBody.get("result").toString();
+
+      String expectedResult = expectedResponse.get("result").toString();
+      String actualResult = responseBody.get("result").toString();
+
+      if (method.equals(Method.TRACE_CALL_MANY)) {
+        // TODO: There are differences in gas cost (causing different balances as well). These are
+        // caused by
+        // OpenEthereum not implementing "Istanbul" correctly. Specially the SSTORE to dirty storage
+        // Slots should cost 800 gas rather than 5000
+        // This should be fixed, e.g. by using
+        // Erigon to
+        // create the expected output, or by making OpenEthereum work correctly.
+        switch (traceType) {
+          case TRACE:
+            expectedResult = filterStringTrace(expectedResult);
+            actualResult = filterStringTrace(actualResult);
+            break;
+          case VM_TRACE:
+            expectedResult = filterStringVmTrace(expectedResult);
+            actualResult = filterStringVmTrace(actualResult);
+            break;
+          case STATE_DIFF:
+            expectedResult = filterStringStateDiff(expectedResult);
+            actualResult = filterStringStateDiff(actualResult);
+            break;
+          default:
+            throw new RuntimeException(
+                "Unrecognized trace type (expected trace | vmTrace | stateDiff)");
+        }
+      }
+
       final ObjectMapper mapper = new ObjectMapper();
       mapper.configure(INDENT_OUTPUT, true);
       assertThat(
               mapper
                   .writerWithDefaultPrettyPrinter()
+                  .withoutAttribute("creationMethod")
                   .writeValueAsString(mapper.readTree(actualResult)))
           .isEqualTo(
               mapper
                   .writerWithDefaultPrettyPrinter()
+                  .withoutAttribute("creationMethod")
                   .writeValueAsString(mapper.readTree(expectedResult)));
     }
 
@@ -181,5 +265,20 @@ public abstract class AbstractJsonRpcHttpBySpecTest extends AbstractJsonRpcHttpS
       final String actualError = responseBody.get("error").toString();
       assertThat(actualError).isEqualToIgnoringWhitespace(expectedError);
     }
+  }
+
+  private String filterStringStateDiff(final String expectedResult) {
+    final Matcher m = GAS_MATCH_FOR_STATE_DIFF.matcher(expectedResult);
+    return m.replaceAll("");
+  }
+
+  private String filterStringVmTrace(final String expectedResult) {
+    final Matcher m = GAS_MATCH_FOR_VM_TRACE.matcher(expectedResult);
+    return m.replaceAll("");
+  }
+
+  private String filterStringTrace(final String expectedResult) {
+    final Matcher m = GAS_MATCH_FOR_TRACE.matcher(expectedResult);
+    return m.replaceAll("");
   }
 }