+++ CHANGELOG.md
@@ -5,9 +5,16 @@
 ### Additions and Improvements
 - Execution layer (The Merge):
   - Execution specific RPC endpoint [[#3378](https://github.com/hyperledger/besu/issues/3378)
+  - Adds JWT authentication to Engine APIs
+- Tracing APIs 
+  - new API methods: trace_rawTransaction, trace_get, trace_callMany
+  - added revertReason to trace APIs including: trace_transaction, trace_get, trace_call, trace_callMany, and trace_rawTransaction
+- Allow mining beneficiary to transition at specific blocks for ibft2 and qbft consensus mechanisms.  [#3115](https://github.com/hyperledger/besu/issues/3115)
+- Return richer information from the PrecompiledContract interface. [\#3546](https://github.com/hyperledger/besu/pull/3546)
 
 ### Bug Fixes
 - Reject locally-sourced transactions below the minimum gas price when not mining. [#3397](https://github.com/hyperledger/besu/pull/3397)
+- Fixed bug with contract address supplied to `debug_accountAt` [#3518](https://github.com/hyperledger/besu/pull/3518)
 
 ## 22.1.1